requirements.txt
fastapi
uvicorn

python-multipart

reportlab
pandas
openpyxl

python-docx
python-pptx
ebooklib
odfpy

pdfminer.six
beautifulsoup4
lxml
PyMuPDF==1.24.4